Minor improvements and GridEmbedding
Some important chnages:
Graph.java: add method "nearestNode" which finds returns the node which is nearest to a given position.
GraphDrawer.java: add new method for drawing graphs with straight edges
GridEmbedding: implements an algorithm which changes the position of nodes in a way such that they lie on a grid
GUI.java: add MouseListener to drawingArea which is allows the user to click on nodes
Node.java: changed the member "int[] position" to "double[] position"

+++ Node.java
@@ -3,36 +3,36 @@ import java.awt.Graphics;
 public class Node {
 
 	private String name;
-	private int[] position;
+	private double[] position;
 
-	Node(String name_, int[] position_) {
+	Node(String name_, double[] position_) {
 		name = name_;
 		position = position_;
 	}
 
 	Node(String name_) {
 		name = name_;
-		position = new int[] {0, 0};
+		position = new double[] {0, 0};
 	}
 
 	public String name() {
 		return name;
 	}
 
-	public int[] position() {
+	public double[] position() {
 		return position;
 	}
 
-	public void setPosition(int x, int y) {
+	public void setPosition(double x, double y) {
 		position[0] = x;
 		position[1] = y;
 	}
 
-	public int x() {
+	public double x() {
 		return position[0];
 	}
 
-	public int y() {
+	public double y() {
 		return position[1];
 	}
