+++ RandomEmbedding.java
@@ -1,10 +1,10 @@
 
 public class RandomEmbedding {
-	
+
 	public static void defineLayout(int width, int height, Graph G) {
-		for(Node node: G.nodes()) {
-			node.setPosition((int) (Math.random() * width), (int) (Math.random() * height));
+		for (Node node : G.nodes()) {
+			node.setPosition(Math.random() * width, Math.random() * height);
 		}
 	};
-	
+
 }
